fix(theme): style waste address dropdowns
- override autocomplete input, placeholder, result list, row and divider colors with semantic theme tokens

- share city and street styles and cover light and dark palette mappings

SVAK-190
