+++ src/components/waste/WasteCityInput.tsx
@@ -1,14 +1,16 @@
 import React, { useContext, useMemo } from 'react';
-import { Platform } from 'react-native';
 import Autocomplete from 'react-native-autocomplete-input';
 
 import { device, normalize, texts } from '../../config';
 import { useFilterCities, useKeyboardHeight, useWasteAddresses } from '../../hooks';
+import { useTheme } from '../../hooks/useTheme';
+import { useThemeStyles } from '../../hooks/useThemeStyles';
 import { OrientationContext } from '../../OrientationProvider';
 import { SettingsContext } from '../../SettingsProvider';
 import { Label } from '../Label';
 import { Wrapper } from '../Wrapper';
-import { useThemeStyles } from '../../hooks/useThemeStyles';
+
+import { createWasteInputStyles } from './wasteInputStyles';
 
 type Props = {
   isFocused: boolean;
@@ -29,7 +31,8 @@ export const WasteCityInput = ({
   setIsFocused,
   setSelectedStreetId
 }: Props) => {
-  const styles = useThemeStyles(createStyles);
+  const { colors } = useTheme();
+  const styles = useThemeStyles(createWasteInputStyles);
   const { dimensions } = useContext(OrientationContext);
   const { globalSettings } = useContext(SettingsContext);
   const { settings = {} } = globalSettings;
@@ -108,58 +111,11 @@ export const WasteCityInput = ({
         }}
         onFocus={() => setIsFocused(true)}
         placeholder={wasteTexts.location}
+        placeholderTextColor={colors.placeholder}
+        selectionColor={colors.primary}
         style={styles.autoCompleteInput}
         value={inputValueCity}
       />
     </Wrapper>
   );
 };
-
-const createStyles = (colors) => ({
-  autoCompleteContainer: {
-    paddingHorizontal: 0
-  },
-
-  autoCompleteInputContainer: {
-    borderColor: colors.gray40,
-    borderRadius: normalize(8),
-    borderWidth: normalize(1),
-    height: normalize(42)
-  },
-
-  autoCompleteInput: {
-    backgroundColor: colors.transparent,
-    color: colors.darkText,
-    paddingLeft: normalize(12),
-    paddingRight: normalize(6),
-    paddingVertical: device.platform === 'ios' ? normalize(10) : normalize(8),
-    fontFamily: 'regular',
-    fontSize: normalize(14),
-    height: normalize(42),
-    lineHeight: normalize(20)
-  },
-
-  autoCompleteList: {
-    paddingHorizontal: normalize(6),
-    position: 'relative',
-    ...Platform.select({
-      ios: {
-        borderWidth: 0
-      },
-      android: {
-        borderColor: colors.gray20,
-        borderRadius: 0,
-        borderWidth: normalize(1),
-        maxHeight: normalize(300)
-      }
-    })
-  },
-
-  autoCompleteListContainer: {
-    elevation: 2,
-    shadowColor: colors.shadow,
-    shadowOffset: { height: 5, width: 0 },
-    shadowOpacity: 0.5,
-    shadowRadius: 3
-  }
-});
