+++ src/components/waste/WasteStreetInput.tsx
@@ -1,15 +1,17 @@
 import React, { useContext, useMemo } from 'react';
-import { Platform } from 'react-native';
 import Autocomplete from 'react-native-autocomplete-input';
 
 import { device, normalize, texts } from '../../config';
 import { useFilterStreets, useKeyboardHeight, useWasteAddresses } from '../../hooks';
+import { useTheme } from '../../hooks/useTheme';
+import { useThemeStyles } from '../../hooks/useThemeStyles';
 import { OrientationContext } from '../../OrientationProvider';
 import { SettingsContext } from '../../SettingsProvider';
 import { Label } from '../Label';
 import { LoadingSpinner } from '../LoadingSpinner';
 import { Wrapper } from '../Wrapper';
-import { useThemeStyles } from '../../hooks/useThemeStyles';
+
+import { createWasteInputStyles } from './wasteInputStyles';
 
 type Props = {
   isFocused: boolean;
@@ -23,7 +25,8 @@ type Props = {
 };
 
 export const WasteStreetInput = ({ isFocused, renderSuggestions, setIsFocused }: Props) => {
-  const styles = useThemeStyles(createStyles);
+  const { colors } = useTheme();
+  const styles = useThemeStyles(createWasteInputStyles);
   const { dimensions } = useContext(OrientationContext);
   const { globalSettings } = useContext(SettingsContext);
   const { settings = {} } = globalSettings;
@@ -90,6 +93,8 @@ export const WasteStreetInput = ({ isFocused, renderSuggestions, setIsFocused }:
         onChangeText={(text) => setInputValue(text)}
         onFocus={() => setIsFocused(true)}
         placeholder={wasteTexts.street}
+        placeholderTextColor={colors.placeholder}
+        selectionColor={colors.primary}
         style={styles.autoCompleteInput}
         value={inputValue}
       />
@@ -97,61 +102,3 @@ export const WasteStreetInput = ({ isFocused, renderSuggestions, setIsFocused }:
     </Wrapper>
   );
 };
-
-const createStyles = (colors) => ({
-  autoCompleteContainer: {
-    paddingHorizontal: 0
-  },
-
-  autoCompleteInputContainer: {
-    borderColor: colors.gray40,
-    borderRadius: normalize(8),
-    borderWidth: normalize(1),
-    height: normalize(42)
-  },
-
-  autoCompleteInput: {
-    backgroundColor: colors.transparent,
-    color: colors.darkText,
-    paddingLeft: normalize(12),
-    paddingRight: normalize(6),
-    paddingVertical: device.platform === 'ios' ? normalize(10) : normalize(8),
-    fontFamily: 'regular',
-    fontSize: normalize(14),
-    height: normalize(42),
-    lineHeight: normalize(20)
-  },
-
-  autoCompleteList: {
-    paddingHorizontal: normalize(6),
-    position: 'relative',
-    ...Platform.select({
-      ios: {
-        borderWidth: 0
-      },
-      android: {
-        borderColor: colors.gray20,
-        borderRadius: 0,
-        borderWidth: normalize(1),
-        maxHeight: normalize(300)
-      }
-    })
-  },
-
-  autoCompleteListContainer: {
-    elevation: 2,
-    shadowColor: colors.shadow,
-    shadowOffset: { height: 5, width: 0 },
-    shadowOpacity: 0.5,
-    shadowRadius: 3
-  },
-
-  noPaddingTop: {
-    paddingTop: 0
-  },
-
-  noBorderTop: {
-    borderTopWidth: 0,
-    marginTop: normalize(-1)
-  }
-});
